scripts/requirements-apply.txt
psycopg[binary]>=3.2
sqlparse>=0.5
python-dotenv>=1.0
